+++ src/antidote/core/__init__.py
@@ -1,3 +1,4 @@
-from .container import DependencyContainer, DependencyInstance, DependencyProvider, Lazy
+from .container import DependencyContainer, DependencyInstance, DependencyProvider
+from antidote.providers.service import LazyFactory
 from .injection import DEPENDENCIES_TYPE, inject
 from .proxy import ProxyContainer