+++ src/antidote/helpers/container.py
@@ -1,11 +1,11 @@
 from ..core import DependencyContainer
-from ..providers import ResourceProvider, ServiceProvider, TagProvider
+from ..providers import LazyCallProvider, ServiceProvider, TagProvider
 
 
 def new_container() -> DependencyContainer:
     container = DependencyContainer()
     container.register_provider(ServiceProvider(container))
-    container.register_provider(ResourceProvider(container))
+    container.register_provider(LazyCallProvider(container))
     container.register_provider(TagProvider(container))
 
     return container